Audio Downloader/requirements.txt
customtkinter==1.0.0
Pillow==9.0.1
pytube==11.0.1
sclib==0.3.0
requests==2.26.0
